refactoring & added reaction model, migration

+++ app/Repositories/CategoryRepository.php
@@ -21,19 +21,45 @@ class CategoryRepository implements CategoryRepositoryInterface
 
     public function store(mixed $data)
     {
-        Category::firstOrCreate(['name' => $data['name']], $data);
+        $category = Category::where('id', $data['id']);
+
+        if ($category) {
+            return ['message' => 'Category already exist!'];
+        }
+
+        Category::create($data);
+
+        return ['message' => 'Category created'];
     }
 
-    public function update(Category $category, $data)
+    public function update(mixed $data)
     {
-        $category->update($data);
+        $category = Category::find($data['id']);
+
+        if ($category) {
+            $category->update($data);
+
+            return ['message' => 'Category updated!'];
+        }
+
+        return ['message' => 'Something went wrong!'];
     }
 
-    public function delete(Category $category)
+    public function delete(int $id)
     {
-        $category->delete();
+        $category = Category::find($id);
+
+        if ($category) {
+            $category->delete();
+
+            return ['message' => 'Category deleted!'];
+        }
+
+        return ['message' => 'Something went wrong!'];
     }
 
+    // TODO закончить реализацию метода получения задач по слагу категорий
+    // перенести в TaskRepo скорее
     public function findOne(string $slug)
     {
         return Category::where('slug', $slug)->firstOrFail();
